save seperately

+++ processing.py
@@ -1,13 +1,12 @@
 import numpy as np
 
 def background_subtracted(data, background):
-    diff = np.divide(np.subtract(data, background, dtype=np.int32), background)
+    diff = np.divide(np.subtract(data, background, dtype=np.int32), background, dtype=np.float64)
     return diff
 
 def float_to_mono(data):
     data = np.clip(data, -1, 1)
     return ((data+1)*32767).astype(np.uint16)
-
 
 
 def common_background(backgrounds):
